refacto: split crypto.ts into focused modules
