+++ frontend/core/src/lib/crypto/encoding.ts
@@ -1,0 +1,31 @@
+/**
+ * base64url <-> bytes helpers (URL- and filesystem-safe, no padding).
+ */
+
+/** Convert ArrayBuffer to base64url string. */
+export function arrayBufferToBase64Url(buffer: ArrayBuffer): string {
+  const bytes = new Uint8Array(buffer);
+  let binary = '';
+  for (let i = 0; i < bytes.byteLength; i++) {
+    binary += String.fromCharCode(bytes[i]);
+  }
+  const base64 = btoa(binary);
+  // Convert to base64url (URL-safe)
+  return base64.replace(/\+/g, '-').replace(/\//g, '_').replace(/=/g, '');
+}
+
+/** Convert base64url string to ArrayBuffer. */
+export function base64UrlToArrayBuffer(base64url: string): ArrayBuffer {
+  // Convert base64url to base64
+  let base64 = base64url.replace(/-/g, '+').replace(/_/g, '/');
+  // Add padding if needed
+  while (base64.length % 4) {
+    base64 += '=';
+  }
+  const binary = atob(base64);
+  const bytes = new Uint8Array(binary.length);
+  for (let i = 0; i < binary.length; i++) {
+    bytes[i] = binary.charCodeAt(i);
+  }
+  return bytes.buffer;
+}
